refactor guide script
